build: clean-up legacy saucelabs karma configuration setup
Cleans up the legacy Saucelabs karma configuration setup.
